flask-ai-service/requirements.txt
flask
flask-cors
chromadb
pypdf2
requests
python-dotenv
google-genai
